fix: Some memory issues with the destruction of Scene objects

+++ Magma/src/Editor/Project/AssetEditorPanel.cpp
@@ -52,7 +52,7 @@ static void EditAsset(Asset asset) {
 	}
 
 	ImGui::Text("Type: %s", type.c_str());
-	ImGui::Text("ID: %llu", asset.ID);
+	ImGui::Text("ID: %lu", (uint64_t)asset.ID);
 }
 
 void AssetEditorPanel::Draw() {
